requirements.txt
torch>=2.2.0
torchvision>=0.17.0
tifffile>=2023.8.12
imagecodecs>=2023.9.18
numpy>=1.24
pyyaml>=6.0
scikit-learn>=1.3
tqdm>=4.66
matplotlib>=3.8
rich>=13.7
seaborn>=0.13
scipy>=1.10
